fix(guard): preserve relative MCP launch paths
Keep directory-qualified launchers such as ./servers/mcp, split shim
PATH tests under the line cap, and treat an empty tools catalog as
empty rather than a listing failure.
